+++ bookmarks/api/serializers.py
@@ -80,6 +80,7 @@ class BookmarkSerializer(serializers.ModelSerializer):
             "is_archived",
             "unread",
             "shared",
+            "sensitive",
             "tag_names",
             "date_added",
             "date_modified",