+++ bookmarks/views/contexts.py
@@ -93,6 +93,21 @@ class ActiveBookmarksContext(RequestContext):
         )
 
 
+class SensitiveBookmarksContext(RequestContext):
+    index_view = "linkding:bookmarks.sensitive"
+    action_view = "linkding:bookmarks.sensitive.action"
+
+    def get_bookmark_query_set(self, search: BookmarkSearch):
+        return queries.query_sensitive_bookmarks(
+            self.request.user, self.request.user_profile, search
+        )
+
+    def get_tag_query_set(self, search: BookmarkSearch):
+        return queries.query_sensitive_bookmark_tags(
+            self.request.user, self.request.user_profile, search
+        )
+
+
 class ArchivedBookmarksContext(RequestContext):
     index_view = "linkding:bookmarks.archived"
     action_view = "linkding:bookmarks.archived.action"
@@ -104,6 +119,21 @@ class ArchivedBookmarksContext(RequestContext):
 
     def get_tag_query_set(self, search: BookmarkSearch):
         return queries.query_archived_bookmark_tags(
+            self.request.user, self.request.user_profile, search
+        )
+
+
+class SensitiveArchivedBookmarksContext(RequestContext):
+    index_view = "linkding:bookmarks.archived.sensitive"
+    action_view = "linkding:bookmarks.archived.sensitive.action"
+
+    def get_bookmark_query_set(self, search: BookmarkSearch):
+        return queries.query_sensitive_archived_bookmarks(
+            self.request.user, self.request.user_profile, search
+        )
+
+    def get_tag_query_set(self, search: BookmarkSearch):
+        return queries.query_sensitive_archived_bookmark_tags(
             self.request.user, self.request.user_profile, search
         )
 
@@ -165,6 +195,7 @@ class BookmarkItem:
         self.favicon_file = bookmark.favicon_file
         self.preview_image_file = bookmark.preview_image_file
         self.is_archived = bookmark.is_archived
+        self.sensitive = bookmark.sensitive
         self.unread = bookmark.unread
         self.owner = bookmark.owner
         self.details_url = context.details(bookmark.id)
@@ -174,6 +205,8 @@ class BookmarkItem:
             css_classes.append("unread")
         if bookmark.shared:
             css_classes.append("shared")
+        if bookmark.sensitive:
+            css_classes.append("sensitive")
 
         self.css_classes = " ".join(css_classes)
 
@@ -278,6 +311,22 @@ class ArchivedBookmarkListContext(BookmarkListContext):
     bulk_edit_enabled = True
     bulk_edit_disabled_actions = "bulk_archive"
     request_context = ArchivedBookmarksContext
+
+
+class SensitiveArchivedBookmarkListContext(BookmarkListContext):
+    list_title = "Sensitive archived bookmarks"
+    search_mode = "archived-sensitive"
+    bulk_edit_enabled = True
+    bulk_edit_disabled_actions = "bulk_archive"
+    request_context = SensitiveArchivedBookmarksContext
+
+
+class SensitiveBookmarkListContext(BookmarkListContext):
+    list_title = "Sensitive bookmarks"
+    search_mode = "sensitive"
+    bulk_edit_enabled = True
+    bulk_edit_disabled_actions = "bulk_unarchive"
+    request_context = SensitiveBookmarksContext
 
 
 class SharedBookmarkListContext(BookmarkListContext):
@@ -513,8 +562,30 @@ class ActiveTagCloudContext(TagCloudContext):
         )
 
 
+class SensitiveTagCloudContext(TagCloudContext):
+    request_context = SensitiveBookmarksContext
+
+    def get_selected_tags(self):
+        return list(
+            queries.get_tags_for_query(
+                self.request.user, self.request.user_profile, self.search.q
+            )
+        )
+
+
 class ArchivedTagCloudContext(TagCloudContext):
     request_context = ArchivedBookmarksContext
+
+    def get_selected_tags(self):
+        return list(
+            queries.get_tags_for_query(
+                self.request.user, self.request.user_profile, self.search.q
+            )
+        )
+
+
+class SensitiveArchivedTagCloudContext(TagCloudContext):
+    request_context = SensitiveArchivedBookmarksContext
 
     def get_selected_tags(self):
         return list(
@@ -617,8 +688,16 @@ class ActiveBookmarkDetailsContext(BookmarkDetailsContext):
     request_context = ActiveBookmarksContext
 
 
+class SensitiveBookmarkDetailsContext(BookmarkDetailsContext):
+    request_context = SensitiveBookmarksContext
+
+
 class ArchivedBookmarkDetailsContext(BookmarkDetailsContext):
     request_context = ArchivedBookmarksContext
+
+
+class SensitiveArchivedBookmarkDetailsContext(BookmarkDetailsContext):
+    request_context = SensitiveArchivedBookmarksContext
 
 
 class SharedBookmarkDetailsContext(BookmarkDetailsContext):